refactor(user-store): 파일 이름 변경한다 (#282)
