+++ src/hooks/store/useUserStore.ts
@@ -1,10 +1,10 @@
 // https://docs.pmnd.rs/zustand/guides/typescript#using-a-vanilla-store-as-a-bound-store
 
-import type { UserState, UserActions } from '@/store/userStore';
+import type { UserState, UserActions } from '@/stores/user-store';
 
 import { useStore } from 'zustand';
 
-import { userStore } from '@/store/userStore';
+import { userStore } from '@/stores/user-store';
 
 export const useUserStore = <T>(
   selector: (state: UserState & UserActions) => T,
